abandom using quadrant information
this mainly because the long arcs will span several quadrant and it is just a liitle bit hard to specify its span. And it may not work at all, for example, you have a arc from 1-4, then you have another short arc in 4, they can be connected. An controversial example, you have an arc 1-3, another arc 3, they are not from an ellipse, so they cannoot be connected. so simply judging the span quadrants cannot indicate whether or not they can be connected.